Add spinner
Rename  context directory to contexts, and rename  the styled  files to style. Also move styled-components to a folder called styled Components and renamed it to index.jsx.

Create reusable components in styled components.

Changed the structure of file and directory.

Applying the theme properties to the entire file

Fixed aside with sticky position